requirements.txt
args==0.1.0
astroid==2.4.2
autopep8==1.5.4
certifi==2020.4.5.1
chardet==3.0.4
clint==0.5.1
colorama==0.4.3
Django==2.2.2
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.9
isort==5.4.2
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
monitor-internet-connection==1.2.1
Pillow==7.1.2
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
zipfile36==0.1.3
